requirements.txt
# Bio Clothes — install with: pip install -r requirements.txt
Flask>=3.0.0
python-dotenv>=1.0.0
